Added health determination for adventurer, switched maze from async t…

+++ MazeGameDomain/Models/Adventurer.cs
@@ -5,7 +5,7 @@ namespace MazeGameDomain.Models
     public class Adventurer
     {
         public string Name { get; set; } = string.Empty;
-        public int Health { get; set; } 
+        public decimal Health { get; set; } 
         public int Class { get; set; }
         public Specialisation Specialisation { get; set; }
         public ICollection<string> Inventory { get; set; } = new List<string>();
